Prevent crash and display server error
Reddit's api should be returning a wallpaper but for some reason it isn't
I have checked other apps made by other people to see if they are experiencing
the same issue and it looks like they are. In the meantime presenting the user
with a message is better than crashing so thats what this commit does.